refactor: put old stuff in deprecated folder
