+++ libs/plugin-generator/src/generators/deprecated-as-reference/all-libraries/schema.d.ts
@@ -1,4 +1,4 @@
-import { GeneratorScope } from '../generator.model';
+import { GeneratorScope } from '../../generator.model';
 
 export interface AllLibrariesGeneratorSchema {
   domain: string;
