+++ libs/plugin-generator/src/generators/deprecated-as-reference/all-libraries/schema.json
