+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library-advanced/files/README.md__template__
