+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library-advanced/files/jest.config.ts__template__
