+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library-advanced/files/src/test-setup.ts__template__
