+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library-advanced/generator.spec.ts
