+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library-advanced/schema.json
