+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/data-access/README.md__template__
