+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/data-access/jest.config.ts__template__
