+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/data-access/src/index.ts__template__
