+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/data-access/src/lib/__fileName__.service.spec.ts__template__
