+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/data-access/src/test-setup.ts__template__
