+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/README.md__template__
