+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/jest.config.ts__template__
