+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__fileName__-add/__fileName__-add.component.html__template__
