+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__fileName__-delete/__fileName__-delete.component.spec.ts__template__
