+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__fileName__-detail/__fileName__-detail.component.spec.ts__template__
