+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__fileName__-edit/__fileName__-edit.component.ts__template__
