+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__fileName__-list/__fileName__-list.component.spec.ts__template__
