+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/__scopeFileName__-__domainFileName__-feature.module.ts__template__
