+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/feature/src/lib/lib.routes.ts__template__
