+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/generator.spec.ts
