+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/schema.json
