+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/ui/README.md__template__
