+++ libs/plugin-generator/src/generators/deprecated-as-reference/angular-library/ui/jest.config.ts__template__
