+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/.eslintrc.json__template__
