+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/README.md__template__
