+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/jest.config.ts__template__
