+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/project.json__template__
