+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/src/index.ts__template__
