+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/src/lib/__optionalFileName____type__.ts__template__
