+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/tsconfig.json__template__
