+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/tsconfig.lib.json__template__
