+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/files/tsconfig.spec.json__template__
