+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/generator.spec.ts
