+++ libs/plugin-generator/src/generators/deprecated-as-reference/typescript-library/schema.d.ts
@@ -1,4 +1,4 @@
-import { GeneratorScope, TypescriptGeneratorType } from './../generator.model';
+import { GeneratorScope, TypescriptGeneratorType } from '../../generator.model';
 
 export interface TypescriptLibraryGeneratorSchema {
   type: TypescriptGeneratorType;
